add depth full frame visualization

+++ WEB-INF/src/developer/depth/XtionRead.java
@@ -1,11 +1,13 @@
 package developer.depth;
 
 /*
- * this class to be run standalone as jar under oculusPrime/xtionread/
+ * this class to be run standalone as xtion.jar under oculusPrime/xtionread/
  * also copy contents of 'OpenNI-Linux-x86-2.2/Redist' folder to oculusPrime/xtionread/
  * called by OpenNIRead.startDepthCam()
+ * (cut/paste this into its own eclipse project, with redist libs in paths)
  */
 
+import java.io.File;
 import java.io.FileOutputStream;
 import java.nio.ByteBuffer;
 import java.nio.ByteOrder;
@@ -24,6 +26,10 @@ public class XtionRead {
     }
 
     public static void main(String s[]) {
+    	
+		File lockfile = new File("/run/shm/xtion.raw.lock");
+		lockfile.delete();
+    	
         // initialize OpenNI
         OpenNI.initialize();
         
@@ -41,27 +47,25 @@ public class XtionRead {
 		mVideoStream.setVideoMode(mode);
 		
 		mVideoStream.start();
-		XtionRead blob = new XtionRead();
+//		XtionRead blob = new XtionRead();
 
+		
 		try {
 
-//        	long lasttime = System.currentTimeMillis();
-//        	int fps = 0;
 	    	
 	    	while (true) {
+	    		
 		    	mLastFrame = mVideoStream.readFrame();
 		    	ByteBuffer frameData = mLastFrame.getData().order(ByteOrder.LITTLE_ENDIAN);
-		    	FileChannel out = new FileOutputStream("/run/shm/xtion.raw").getChannel();
-		    	out.write(frameData);
-		    	out.close();
- 
-//		    	fps ++;
-//		    	long now = System.currentTimeMillis();
-//		    	if (now - lasttime  >= 1000) {
-//		    		System.out.println(fps);
-//		    		lasttime = now;
-//		    		fps = 0;
-//		    	}
+	    		if (!lockfile.exists()) {
+	    			lockfile.createNewFile();
+			    	FileChannel out = new FileOutputStream("/run/shm/xtion.raw").getChannel();
+	
+			    	out.write(frameData);
+			    	out.close();
+			    	lockfile.delete();
+//			    	Thread.sleep(100);
+	    		}
 		    	
 	    	}
